requirements.txt
alembic==0.9.2
amqp==1.4.9
aniso8601==1.2.1
anyjson==0.3.3
appdirs==1.4.3
auth0-python==3.0.0
billiard==3.3.0.23
boto==2.47.0
celery==3.1.25
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
click==6.7
cov-core==1.15.0
coverage==4.4.1
dotenv==0.0.5
filechunkio==1.8
Flask==0.12.2
Flask-Celery==2.4.3
Flask-Cors==3.0.2
Flask-Markdown==0.3
Flask-Migrate==2.0.4
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
kombu==3.0.37
Mako==1.0.6
Markdown==2.6.8
MarkupSafe==1.0
meld3==1.0.2
nose2==0.6.5
packaging==16.8
psycopg2==2.7.1
pycparser==2.17
pyparsing==2.2.0
python-dateutil==2.6.0
python-dotenv==0.6.3
python-editor==1.0.3
pytz==2017.2
requests==2.17.3
six==1.10.0
SQLAlchemy==1.1.10
supervisor==3.3.2
urllib3==1.21.1
vine==1.1.3
Werkzeug==0.12.2
